refactor(card): component themes
