--- src/components/card/card.actions.bootstrap.scss
@@ -1,7 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-card-actions {
-        padding: utils.rem(16px);
-    }
-}
