--- src/components/card/card.bootstrap.scss
@@ -1,24 +1,0 @@
-@use '../../styles/utilities' as utils;
-@use 'card.header.bootstrap' as card-header;
-@use 'card.content.bootstrap' as card-content;
-@use 'card.actions.bootstrap' as card-actions;
-
-@mixin theme() {
-    igc-card {
-        background: #{utils.color(surface, 500)};
-        border-radius: #{utils.rem(4px)};
-    }
-
-    igc-card[elevated] {
-        box-shadow: var(--igc-elevation-2);
-
-        &:hover,
-        &:focus-within {
-            box-shadow: var(--igc-elevation-2);
-        }
-    }
-
-    @include card-header.theme();
-    @include card-content.theme();
-    @include card-actions.theme();
-}
