--- src/components/card/card.content.bootstrap.scss
@@ -1,15 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-card-content {
-        @include utils.type-category('body-1');
-
-        color: #{utils.color(gray, 900)};
-
-        > * {
-            @include utils.type-category('body-1');
-
-            opacity: 1;
-        }
-    }
-}
