--- src/components/card/card.content.indigo.scss
@@ -1,11 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-card-content {
-        color: #{utils.color(gray, 700)};
-
-        > * {
-            opacity: 1;
-        }
-    }
-}
