--- src/components/card/card.header.bootstrap.scss
@@ -1,15 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-card-header::part(title) {
-        @include utils.type-category('h5');
-
-        color: #{utils.color(gray, 900)};
-        opacity: 1;
-    }
-
-    igc-card-header::part(subtitle) {
-        color: #{utils.color(gray, 800)};
-        opacity: 1;
-    }
-}
