--- src/components/card/card.header.indigo.scss
@@ -1,13 +1,0 @@
-@use '../../styles/utilities' as utils;
-
-@mixin theme() {
-    igc-card-header::part(title) {
-        color: #{utils.color(gray, 900)};
-        opacity: 1;
-    }
-
-    igc-card-header::part(subtitle) {
-        color: #{utils.color(gray, 500)};
-        opacity: 1;
-    }
-}
