+++ src/components/card/themes/light/card.actions.bootstrap.scss
@@ -1,0 +1,6 @@
+@use './card.actions.material.scss';
+@use '../../../../styles/utilities' as *;
+
+:host {
+  padding: rem(16px);
+}
