+++ src/components/card/themes/light/card.bootstrap.scss
@@ -1,0 +1,12 @@
+@use '../../../../styles/utilities' as *;
+@use '../light/card.material.scss';
+
+:host {
+  border-radius: rem(4px);
+}
+
+:host([elevated]),
+:host([elevated]:hover),
+:host([elevated]:focus-within) {
+  box-shadow: var(--igc-elevation-2);
+}
