+++ src/components/card/themes/light/card.content.indigo.scss
@@ -1,0 +1,10 @@
+@use './card.content.material.scss';
+@use '../../../../styles/utilities' as *;
+
+:host {
+  color: color(gray, 700);
+}
+
+::slotted(*) {
+  opacity: 1;
+}
