+++ src/components/card/themes/light/card.fluent.scss
@@ -1,0 +1,20 @@
+@use '../../../../styles/utilities' as *;
+@use '../light/card.material.scss';
+
+:host {
+  border: rem(1px) solid color(gray, 100);
+  border-radius: 0;
+}
+
+:host([elevated]:hover),
+:host([elevated]:focus-within) {
+  &::after {
+    position: absolute;
+    content: '';
+    width: 100%;
+    height: 100%;
+    transition: box-shadow 0.1s ease-out;
+    box-shadow: inset 0 0 0 rem(2px) color(gray, 300);
+    pointer-events: none;
+  }
+}
