+++ src/components/card/themes/light/card.header.bootstrap.scss
@@ -1,0 +1,14 @@
+@use './card.header.material.scss';
+@use '../../../../styles/utilities' as *;
+
+::slotted([slot='title']) {
+  @include type-category('h5');
+
+  color: color(gray, 900);
+  opacity: 1;
+}
+
+::slotted([slot='subtitle']) {
+  color: color(gray, 800);
+  opacity: 1;
+}
