+++ src/components/card/themes/light/card.header.indigo.scss
@@ -1,0 +1,12 @@
+@use './card.header.material.scss';
+@use '../../../../styles/utilities' as *;
+
+::slotted([slot='title']) {
+  color: color(gray, 900);
+  opacity: 1;
+}
+
+::slotted([slot='subtitle']) {
+  color: color(gray, 500);
+  opacity: 1;
+}
