+++ src/styles/themes/dark/fluent.scss
@@ -9,7 +9,6 @@
     $button-color: utils.color(surface, 500),
     $hover-color: utils.color(primary, 300)
 );
-@use '../../../components/card/card.fluent' as card;
 @use '../../../components/checkbox/checkbox.fluent' as checkbox with (
     $border-color: utils.color(gray, 700),
     $hover-color: utils.color(gray, 900),
@@ -62,7 +61,6 @@ $palette: utils.palette(
 
 @include base.root-styles($palette);
 @include button.theme();
-@include card.theme();
 @include checkbox.theme();
 @include radio.theme();
 @include switch.theme();
