+++ src/styles/themes/light/fluent.scss
@@ -2,7 +2,6 @@
 @use '../base/fluent' as base;
 @use '../../../components/button/button.fluent' as button;
 @use '../../../components/button/icon-button.fluent' as icon-button;
-@use '../../../components/card/card.fluent' as card;
 @use '../../../components/checkbox/checkbox.fluent' as checkbox;
 @use '../../../components/radio/radio.fluent' as radio;
 @use '../../../components/list/list.fluent' as list;
@@ -22,7 +21,6 @@ $palette: utils.palette(
 
 @include base.root-styles($palette);
 @include button.theme();
-@include card.theme();
 @include checkbox.theme();
 @include radio.theme();
 @include list.theme();
